+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - Menu for setting up SmartAds networks into the Unity Editor.
 - Menu for downloading Android SmartAds library dependencies.
+- Method to Transaction for setting the receipt signature.
 ### Fixed
 - Compilation errors in SmartAds when missing iOS build support.
 ### Removed